+++ fdbclient/ManagementAPI.actor.cpp
@@ -65,7 +65,7 @@ std::map<std::string, std::string> configForToken( std::string const& mode ) {
 		std::string key = mode.substr(0, pos);
 		std::string value = mode.substr(pos+1);
 
-		if( (key == "logs" || key == "proxies" || key == "resolvers" || key == "remote_logs" || key == "satellite_logs") && isInteger(value) ) {
+		if( (key == "logs" || key == "proxies" || key == "resolvers" || key == "remote_logs" || key == "satellite_logs" || key == "usable_regions") && isInteger(value) ) {
 			out[p+key] = value;
 		}
 
@@ -135,8 +135,7 @@ std::map<std::string, std::string> configForToken( std::string const& mode ) {
 	} else
 		redundancySpecified = false;
 	if (redundancySpecified) {
-		out[p+"storage_replicas"] =
-			out[p+"storage_quorum"] = redundancy;
+		out[p+"storage_replicas"] = redundancy;
 		out[p+"log_replicas"] = log_replicas;
 		out[p+"log_anti_quorum"] = "0";
 
@@ -153,7 +152,7 @@ std::map<std::string, std::string> configForToken( std::string const& mode ) {
 	std::string remote_redundancy, remote_log_replicas;
 	IRepPolicyRef remoteTLogPolicy;
 	bool remoteRedundancySpecified = true;
-	if (mode == "remote_none") {
+	if (mode == "remote_default") {
 		remote_redundancy="0";
 		remote_log_replicas="0";
 		remoteTLogPolicy = IRepPolicyRef();
@@ -244,7 +243,6 @@ bool isCompleteConfiguration( std::map<std::string, std::string> const& options 
 
 	return 	options.count( p+"log_replicas" ) == 1 &&
 			options.count( p+"log_anti_quorum" ) == 1 &&
-			options.count( p+"storage_quorum" ) == 1 &&
 			options.count( p+"storage_replicas" ) == 1 &&
 			options.count( p+"log_engine" ) == 1 &&
 			options.count( p+"storage_engine" ) == 1;
@@ -1212,7 +1210,7 @@ ACTOR Future<Void> waitForFullReplication( Database cx ) {
 				}
 			}
 
-			if( !watchFutures.size() || (config.remoteTLogReplicationFactor == 0 && watchFutures.size() < config.regions.size())) {
+			if( !watchFutures.size() || (config.usableRegions == 1 && watchFutures.size() < config.regions.size())) {
 				return Void();
 			}